Use arrays instead of slices
This proposes the use of arrays instead of slices
for the `cfft_$N` and `rfft_$N` functions.

Though it is it a backwards incompatible API change,
it essentially moves the length checking to compile
time instead of runtime, as the lengths were limited
to a specific values before as well.

Everyone, who still would want to use the slice method
could go for `CFftN$N::transform(input)` directly,
as this functions takes a slice and checks the length of it anyways.

In case of the `cfft` functions, we can remove any call, which panics,
as the slice / array is changed in place and can be return as is.

Though for the `rfft` case, we have to call `try_into().unwrap()`.
I can not think of another way, in which the compiler could check
the correct slice length, so that it can be converted into an array
safely without panic.

Though theoretically this function shouldn't panic as well.

Therefor I removed any mentions of panics in the documentation as well.