Add coverage
This commit adds support to collect the test coverage.
It uses cargo-tarpaulin.
The v0.16 prebuilt binary is downloaded from the github release
to reduce built times, as cargo-subcommands are heavy on compile times.

The `.gitlab-ci.yml` is modified to run cargo-tarpaulin for the
`rust-latest` build. The `coverage` key is used to grep the current
coverage percentage, which gitlab will show in it's overlay
for any commit, where the CI was run.